create service

+++ src/main/java/com/example/springpractice/repository/AnimeRepository.java
@@ -3,5 +3,5 @@ package com.example.springpractice.repository;
 import com.example.springpractice.models.Anime;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnimeRepository extends JpaRepository<Anime, Long> {
+public interface AnimeRepository extends JpaRepository<Anime, Integer> {
 }
